refacotr embed
