--- openproblems/tasks/_batch_integration/batch_integration_embed/metrics/_utils.py
@@ -1,5 +1,0 @@
-def _get_split(adata):
-    uni = adata
-    uni.obsm["X_pca"] = uni.obsm["X_uni_pca"]
-    uni.X = uni.layers["log_normalized"]
-    return (uni, adata)
