+++ openproblems/tasks/_batch_integration/batch_integration_embed/metrics/graph_connectivity.py
@@ -1,5 +1,6 @@
 from .....tools.decorators import metric
 from ...batch_integration_graph import metrics as graph_metrics
+from .utils import embedding_to_graph
 
 """
 The graph connectivity metric assesses whether the kNN graph representation,
@@ -22,8 +23,4 @@ on the kNN graph, it can be used to evaluate all integration outputs.
 
 @metric(**graph_metrics.graph_connectivity.metadata)
 def graph_connectivity(adata):
-    from scanpy.pp import neighbors
-
-    if not (adata.uns["is_baseline"] and "neighbors" in adata.uns):
-        neighbors(adata, use_rep="X_emb")
-    return graph_metrics.graph_connectivity(adata)
+    return graph_metrics.graph_connectivity(embedding_to_graph(adata))
