+++ openproblems/tasks/_batch_integration/batch_integration_embed/metrics/nmi.py
@@ -1,5 +1,6 @@
 from .....tools.decorators import metric
 from ...batch_integration_graph import metrics as graph_metrics
+from .utils import embedding_to_graph
 
 """NMI compares the overlap of two clusterings.
 We used NMI to compare the cell-type labels with Louvain clusters computed on
@@ -15,8 +16,4 @@ the scikit-learn27 (v.0.22.1) implementation of NMI.
 
 @metric(**graph_metrics.nmi.metadata)
 def nmi(adata):
-    from scanpy.pp import neighbors
-
-    if not (adata.uns["is_baseline"] and "neighbors" in adata.uns):
-        neighbors(adata, use_rep="X_emb")
-    return graph_metrics.nmi(adata)
+    return graph_metrics.nmi(embedding_to_graph(adata))
