+++ openproblems/tasks/_batch_integration/batch_integration_feature/metrics/utils.py
@@ -1,3 +1,6 @@
+from ...batch_integration_embed.metrics.utils import embedding_to_graph
+
+
 def feature_to_embedding(adata):
     import scanpy as sc
 
@@ -6,17 +9,6 @@ def feature_to_embedding(adata):
         return adata
 
     adata.obsm["X_emb"] = sc.pp.pca(adata.X)
-    return adata
-
-
-def embedding_to_graph(adata):
-    import scanpy as sc
-
-    if adata.uns["is_baseline"] and "neighbors" in adata.uns:
-        # precomputed; do nothing
-        return adata
-
-    sc.pp.neighbors(adata, use_rep="X_emb")
     return adata
